#17: moved and renamed redux types part I
