--- src/redux/types.js
@@ -1,29 +1,0 @@
-'use strict';
-
-/**
- * Action to add an item to the list.
- *
- * @type {string}
- */
-export const SHOP_LIST_ADD_ITEM = 'SHOP_LIST_ADD_ITEM';
-
-/**
- * Action to update an item in the list.
- *
- * @type {string}
- */
-export const SHOP_LIST_UPDATE_ITEM = 'SHOP_LIST_UPDATE_ITEM';
-
-/**
- * Action to remove item from list.
- *
- * @type {string}
- */
-export const SHOP_LIST_REMOVE_ITEM = 'SHOP_LIST_REMOVE_ITEM';
-
-/**
- * Action to edit item from list.
- *
- * @type {string}
- */
-export const SHOP_LIST_EDIT_ITEM = 'SHOP_LIST_EDIT_ITEM'; // TODO: rename, it's loading to Editor => AddItem component should be renamed to ItemEditor
