+++ src/redux/types/shop/editor.js
@@ -1,0 +1,15 @@
+'use strict';
+
+/**
+ * Action to add an item from editor to the list.
+ *
+ * @type {string}
+ */
+export const SHOP_EDITOR_ITEM_ADD = 'SHOP_EDITOR_ITEM_ADD';
+
+/**
+ * Action to update an item from the editor in the list.
+ *
+ * @type {string}
+ */
+export const SHOP_EDITOR_ITEM_UPDATE = 'SHOP_EDITOR_ITEM_UPDATE';
