+++ src/redux/types/shop/list.js
@@ -1,0 +1,15 @@
+'use strict';
+
+/**
+ * Action to remove item from list.
+ *
+ * @type {string}
+ */
+export const SHOP_LIST_REMOVE_ITEM = 'SHOP_LIST_REMOVE_ITEM';
+
+/**
+ * Action to edit item from list (loading it into the editor).
+ *
+ * @type {string}
+ */
+export const SHOP_LIST_EDIT_ITEM = 'SHOP_LIST_EDIT_ITEM';
